requirements.txt
psycopg2-binary
requests
# Add any other dependencies here
